move load logic to movie view and remove redux-thunk
