+++ frontend/src/index.js
@@ -2,16 +2,15 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import { Provider } from 'react-redux';
-import { createStore, compose, applyMiddleware } from 'redux';
+import { createStore, compose } from 'redux';
 import { BrowserRouter } from 'react-router-dom';
 import rootReducers from './services/rootReducers';
-import thunk from 'redux-thunk';
 
 import App from './App';
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
-const store = createStore(rootReducers, applyMiddleware(thunk));
+const store = createStore(rootReducers, composeEnhancers());
 
 ReactDOM.render(
     <BrowserRouter>
